simplify client
