--- node/service/src/client/mod.rs
@@ -1,5 +1,0 @@
-pub mod crab;
-pub use crab::CrabClient;
-
-pub mod darwinia;
-pub use darwinia::DarwiniaClient;
